power: smb5-lib: Limit ICL for Audio Adapter
The USBC spec leaves remove for an Audio Adapter that also
has a Charging USBC port.  The spec also limits the power
from that port to 500 mA.  Ensure that when typec_mode is
Audio the Charger pulls not more then 500 mA by setting ICL.

Also because type_mode could be None but still have a Audio
adapter with Charger use VBUS presence as a indication
that 500 mA limit shoul dbe used to charge as well as
setting USB present and online.

This change will have the effect of allowing any cable that
lacks Rp to charge the Phone at 500 mA.  There is no
way to determine the Source is an Audio adapter without
an 3.5 mm headset versus a Raw VBUS applied to the Port.
If future revision of the USB Type C spec are modified then
this must be as well.

refer to https://gerrit.mot.com/1017220

Change-Id: I4451cf753f7c7877d06fd939225b52832e13a3fb
Signed-off-by: Haijian Ma <mahj8@motorola.com>
Reviewed-on: https://gerrit.mot.com/1274028
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key