msm: fastrpc: add debug logs for adsp client panic
- add error code, Distinguish the type of panic;
- Print TGID when adsp client panic;
Currently, when adsp slient panic occurs, if the root cause of the adsp panic is caused
by adsp, we can use the adsp's crash and locate the root-cause.
But, If the ADSP crash is caused by an abnormality (similar to no response, etc.) on the AP
side, the current method can only obtain the full ramdump then locate it. It is difficult.
so, qcom recommends add some necessary error info output in kernel'logs when adsp panic.

Change-Id: Ifea78b8e12929c8ecbea42f2c0b02d279200dccb
Signed-off-by: huangjq9 <huangjq9@motorola.com>
Reviewed-on: https://gerrit.mot.com/1820377
SME-Granted: SME Approvals Granted
SLTApproved: Slta Waiver
Tested-by: Jira Key
Reviewed-by: Jun Weng <wengjun1@motorola.com>
Reviewed-by: Huosheng Liao <liaohs@motorola.com>
Submit-Approved: Jira Key